requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile setup.py
#
appdirs==1.4.3            # via deckz (setup.py)
cached-property==1.5.1    # via pygit2
cffi==1.14.0              # via pygit2
click==7.1.1              # via deckz (setup.py)
coloredlogs==14.0         # via deckz (setup.py)
humanfriendly==8.1        # via coloredlogs
jinja2==2.11.1            # via deckz (setup.py)
markupsafe==1.1.1         # via jinja2
pycparser==2.20           # via cffi
pygit2==1.1.1             # via deckz (setup.py)
pyyaml==5.3.1             # via deckz (setup.py)
